requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
pytest-cov
python-dotenv
